+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_14/level_1_2/FruitStorage.java
@@ -1,4 +1,4 @@
-package student_aleksandrs_korsaks.ak_lesson_14.level_1;
+package student_aleksandrs_korsaks.ak_lesson_14.level_1_2;
 
 import java.util.ArrayList;
 import java.util.List;
